--- a/src/bartholomew.rs
+++ b/src/bartholomew.rs
@@ -1,13 +1,23 @@
-use crate::content;
+use crate::content::{self, all_pages_load};
 use crate::response::{self, *};
 use crate::template::{self, DynamicTemplateParams};
 use anyhow::{anyhow, Result};
+use chrono::{DateTime, Utc};
+use serde::{Deserialize, Serialize};
 use spin_sdk::{
     http::{Request, Response},
     http_component,
+    key_value::Store,
 };
 use std::collections::HashMap;
 use std::path::PathBuf;
+use std::str::FromStr;
+
+#[derive(Serialize, Deserialize)]
+struct RenderedPageCacheEntry {
+    expiry_time: DateTime<Utc>,
+    content: String,
+}
 
 /// The entry point to Bartholomew.
 #[http_component]
@@ -39,7 +49,26 @@ pub fn render(req: Request) -> Result<Response> {
         }
         None => DEFAULT_INDEX.to_owned(),
     };
+
     eprintln!("Request path: {}", &path_info);
+    let store = Store::open_default()?;
+
+    if !disable_cache {
+        let cached_content = store.get(path_info.clone());
+        if let Ok(rendered) = cached_content {
+            let data: RenderedPageCacheEntry = serde_json::from_slice(&rendered)?;
+            if Utc::now() < data.expiry_time {
+                let content_encoding = req.headers().get(http::header::ACCEPT_ENCODING);
+                return response::send_result(
+                    path_info,
+                    data.content,
+                    DEFAULT_CONTENT_TYPE.to_owned(),
+                    content_encoding,
+                    None,
+                );
+            }
+        }
+    }
 
     // Load the site config.
     let mut config: template::SiteInfo = toml::from_slice(&std::fs::read(CONFIG_FILE)?)?;
@@ -140,10 +169,16 @@ pub fn render(req: Request) -> Result<Response> {
                     if let Some(val) = dynamic_template_params {
                         doc.head.template = Some(val.dynamic_template_name.to_owned());
                     }
+                    let disable_page_cache = doc.head.disable_page_cache;
 
                     let data = engine
                         .render_template(doc, config, req.headers().to_owned())
                         .map_err(|e| anyhow!("Rendering {:?}: {}", &content_path, e))?;
+
+                    // If global and page specific caching has not been turned off by default, cache rendered content
+                    if !disable_cache && !disable_page_cache.unwrap_or_default() {
+                        cache_rendered_page(&store, &path_info, data.clone())?;
+                    }
 
                     let content_encoding = req.headers().get(http::header::ACCEPT_ENCODING);
 
@@ -157,4 +192,17 @@ pub fn render(req: Request) -> Result<Response> {
             response::not_found(path_info, body)
         }
     }
+}
+
+// This function caches the rendered content
+// along with the expiry time in the cache
+// unless the page has been opted out
+fn cache_rendered_page(store: &Store, path_info: &str, rendered_content: String) -> Result<()> {
+    let index_cache = all_pages_load(PathBuf::from_str(CONTENT_PATH)?, false)?;
+    let rendered_cache = RenderedPageCacheEntry {
+        expiry_time: index_cache.cache_expiration,
+        content: rendered_content,
+    };
+    store.set(path_info, serde_json::to_string(&rendered_cache)?)?;
+    Ok(())
 }