without caching in kv
```bash
bombardier -c 1 -n 1000 http://localhost:3000/cloud/data-redis
Bombarding http://localhost:3000/cloud/data-redis with 1000 request(s) using 1 connection(s)
 1000 / 1000 [=========================================================] 100.00% 68/s 14s
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec        69.01      24.34     104.92
  Latency       14.48ms     1.66ms    25.97ms
  HTTP codes:
    1xx - 0, 2xx - 1000, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:     2.20MB/s
```

with caching 
```bash
bombardier -c 1 -n 1000 http://localhost:3000/cloud/data-postgres
Bombarding http://localhost:3000/cloud/data-postgres with 1000 request(s) using 1 connection(s)
 1000 / 1000 [=========================================================] 100.00% 998/s 1s
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec      1168.36     367.53    1642.22
  Latency        0.86ms     5.59ms   177.52ms
  HTTP codes:
    1xx - 0, 2xx - 1000, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    34.10MB/s
```